+++ lib/src/extensions/replace_where.dart
@@ -1,6 +1,6 @@
 extension ReplaceWhereExtension<T> on Iterable<T> {
   /// Replaces every element that matches the [where] condition with [value].
-  Iterable<T> replaceWhere(T value, {required bool Function(T) where}) sync* {
+  Iterable<T> replaceWhere(T value, bool Function(T) where) sync* {
     final iterator = this.iterator;
 
     while (iterator.moveNext()) {